refactor: strengthen the CodeBlitz brand
