+++ packages/sumi-core/src/common/enum.ts
@@ -1,4 +1,4 @@
 export enum REPORT_NAME {
-  ALEX_APP_START_ERROR = 'alex.appStartError',
+  APP_START_ERROR = 'alex.appStartError',
   CODE_SERVICE_REQUEST_ERROR = 'code-service.requestError',
 }
