+++ packages/toolkit/webpack/config.build.js
@@ -78,7 +78,6 @@ const globalBundle = createWebpackConfig({
       [config.editorAllGlobalEntry]: './packages/core/src/editor.all',
       [config.editorAllGlobalMiniEntry]: './packages/core/src/editor.all',
     },
-    // 此处 bundle 的包仅作为 commonjs 使用，但因为 external 原因会导致 webpack4 加载 bundle 出错，因此还是使用 umd
     output: {
       library: 'Alex',
       libraryTarget: 'global',
